+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - **[Changed]** streamlined `ParallelState` to prevent potential interrupt race issues. Enable `AtLeastOneState` to overwrite two function for its functionality.
 - **[Changed]** Added `wait_for_internal_states` and `interrupt_internal_states` prototype for nested states to simplify interruptions and running.
+- **[Changed]** Streamlined the status flags such that `UNKNOWN` is set at the beginning of the start method. This allows checking of edge cases between the start of `start` and the start of the execution thread.
 - **[Fixed]** condition in `AtLeastOneState` where other states aren't destroyed when exiting under success condition.
 
 ## [0.3.5] - 2021-06-15